+++ src/__tests__/server.tests.js
@@ -9,17 +9,20 @@ jest.mock('../logger')
 jest.mock('../slack', () => jest.fn())
 
 describe('app', () => {
-  const doRequest = (status = 'Available') =>
-    request(app)
+  const doRequest = (response) => {
+    const { presence_status = 'Available', email } = response || {}
+    return request(app)
       .post('/')
       .send({
         // @see https://marketplace.zoom.us/docs/api-reference/webhook-reference/user-events/presence-status-updated
         payload: {
           object: {
-            presence_status: status,
+            email,
+            presence_status,
           },
         },
       })
+  }
 
   afterEach(() => {
     updateSlackStatus.mockClear()
@@ -36,19 +39,34 @@ describe('app', () => {
     expect(response.status).toStrictEqual(200)
   })
 
-  it.each(['Available', 'Away'])(
-    'updateSlackStatus to "not in a meeting" when presence_status is %s',
-    async (status) => {
-      await doRequest(status)
-      expect(updateSlackStatus).toHaveBeenCalledWith(false)
+  it.each([{ presence_status: 'Available' }, { presence_status: 'Away' }])(
+    'invokes updateSlackStatus to "not in a meeting" with %o',
+    async (response) => {
+      await doRequest(response)
+      expect(updateSlackStatus).toHaveBeenCalledWith({
+        isInMeeting: false,
+      })
     },
   )
 
-  it.each([ZOOM_IN_MEETING_STATUS])(
-    'updateSlackStatus to "in a meeting" when presence_status is %s',
-    async (status) => {
-      await doRequest(status)
-      expect(updateSlackStatus).toHaveBeenCalledWith(true)
+  it.each([{ presence_status: ZOOM_IN_MEETING_STATUS }])(
+    'invokes updateSlackStatus to "in a meeting" with %o',
+    async (response) => {
+      await doRequest(response)
+      expect(updateSlackStatus).toHaveBeenCalledWith({
+        isInMeeting: true,
+      })
+    },
+  )
+
+  it.each([{ email: 'your-address@mail.com' }, { email: '' }])(
+    'invokes updateSlackStatus with proper email with %o',
+    async (response) => {
+      await doRequest(response)
+      expect(updateSlackStatus).toHaveBeenCalledWith({
+        isInMeeting: false,
+        email: response.email,
+      })
     },
   )